raggy_fastapi_deploy/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
transformers
torch
scikit-learn
faiss-cpu